The matches php file
So for this file the main components are the name and view my matches features. When the user presses "View My Matches," the form submits its data as a GET request to matches-submit.php. The name of the query parameter sent should be name, and its value should be the encoded text typed by the user. For example, when the user views matches for Jiang Li,the URL should be: matches-submit.php?name=Jiang+Li