User matches
Read the name from the page's "name" query parameter and finds which other singles match the given user. Output the HTML to display the matches, in the order they were originally found in the file. Each match has the image user.jpg, the person's name, and an unordered list with their gender, age, personality type, and OS.
- $_GET and $_POST are superglobals. They do not need to be passed as a function parameter.
- Here I am also retrieving the user's info from singles.txt file